+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroPluginExtension.kt
@@ -205,8 +205,11 @@ constructor(
       .convention(OptionalBindingBehavior.DEFAULT)
 
   /** Enable/disable automatic transformation of providers to be private. Enabled by default. */
+  @Deprecated(
+    "Transforming providers to private is deprecated as it results in less efficient code generation"
+  )
   public val transformProvidersToPrivate: Property<Boolean> =
-    objects.booleanProperty("metro.transformProvidersToPrivate", true)
+    objects.booleanProperty("metro.transformProvidersToPrivate", false)
 
   /**
    * Configures the Metro compiler plugin to warn, error, or do nothing when it encounters `public`