+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Changelog
     - By extension, this resolves Metro's most subscribed issue ([#460](https://github.com/ZacSweers/metro/issues/460)) by enabling cross-module contribution/aggregation features in non-jvm/android compilations 🎉.
     - Note that there is a separate known kotlinc issue around qualifier annotations that affects native builds and is targeted for `2.3.20-Beta2`. Follow [#1556](https://github.com/ZacSweers/metro/issues/1556) for updates.
 - **[Gradle / FIR]** Enable top-level function injection by default on Kotlin `2.3.20-Beta1` or later.
+- **[Gradle / FIR]** Disable automatic transformation of providers to private + deprecate the option, as this results in less efficient code generation and somewhat unclear error messages. It's recommended to write a lint check in another static analysis tool if you want this, for example [this implementation from slack-lints](https://github.com/slackhq/slack-lints/blob/main/slack-lint-checks/src/main/java/slack/lint/DoNotCallProvidersDetector.kt).
+  - Metro _may_ add an option for the inverse in the future. i.e., allow writing a private provider in source but then transform it to `internal` at compile-time.
 
 ### New
 
@@ -69,6 +71,7 @@ Changelog
 - **[IR]** Don't cache creator-less scoped graph extension impls in their parent graphs. This was initially implemented this way due to a misunderstanding to how Dagger generated subcomponents! Getters for graph extensions now always return new instances.
 - **[IR]** Check for `open` or `final` modality on classes first before searching for injectable constructors.
 - **[IR]** Mark all supertypes' member injections as dependencies of subtypes' member injections. That's a wordier way of saying that member-injecting a class `Dog` that extends `Animal` also depends on member-injected dependencies of `Animal`.
+- **[IR]** Don't process companion objects of graphs or binding containers independently of their parent classes.
 - **[IR/Native]** Work around `Abstract function '___' is not implemented in non-abstract class 'Impl'` info warnings from native compilations.
 - **[IR/Dynamic Graphs]** Always check if a key is replaced by a dynamic type key when building binding graphs. Previously, some bindings would be added even if a dynamic key was available, resulting in unexpected `DuplicateBinding` errors.
 - **[IC]** Record lookups of contributed classes when looking up hints from IR. Previously Metro only recorded a lookup of the generated hint function, which appears to not be enough for Kotlin 2.3.20.